types: refactor RsbuildConfig typings
