+++ packages/compat/webpack/src/types/config/output.ts
@@ -2,14 +2,7 @@ import type {
   SharedOutputConfig,
   NormalizedSharedOutputConfig,
 } from '@rsbuild/shared';
-import type { CopyPluginOptions } from '../thirdParty';
 
-export type OutputConfig = SharedOutputConfig & {
-  /**
-   * Copies the specified file or directory to the dist directory.
-   */
-  copy?: CopyPluginOptions | CopyPluginOptions['patterns'];
-};
+export type OutputConfig = SharedOutputConfig;
 
-export type NormalizedOutputConfig = OutputConfig &
-  NormalizedSharedOutputConfig;
+export type NormalizedOutputConfig = NormalizedSharedOutputConfig;
