+++ packages/core/src/rspack-provider/types/index.ts
@@ -2,4 +2,4 @@ export * from './hooks';
 export * from './config';
 export * from './plugin';
 export * from './context';
-export * from './rspack';
+export * from '@rsbuild/shared/src/types/rspack';
