+++ packages/core/tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "./dist",
     "baseUrl": "./"
   },
-  "include": ["src"]
+  "include": ["src", "../shared/src/types/rspack.ts"]
 }
