+++ packages/shared/src/types/index.ts
@@ -2,6 +2,7 @@ import type WebpackChain from 'webpack-chain';
 
 export type { WebpackChain };
 
+export * from './rspack';
 export * from './rsbuild';
 export * from './context';
 export * from './utils';
