+++ packages/shared/src/types/rspack.ts
